Unified the high DPI handling. All platforms should now be emulating the "retina" model of DPI-invariant window coordinates (that may be fractional) and framebuffer size specified in pixels.

+++ include/mnm/mnm.h
@@ -74,13 +74,13 @@ void quit(void);
 ///
 /// @returns Window width in screen coordinates.
 ///
-int width(void);
+float width(void);
 
 /// Returns window height in screen coordinates.
 ///
 /// @returns Window height in screen coordinates.
 ///
-int height(void);
+float height(void);
 
 /// Returns window aspect ratio, i.e, its width divided by its height.
 ///
@@ -131,27 +131,27 @@ enum
 ///
 /// @returns Mouse X position in screen coordinates.
 ///
-int mouse_x(void);
+float mouse_x(void);
 
 /// Returns mouse Y position in screen coordinates.
 ///
 /// @returns Mouse Y position in screen coordinates.
 ///
-int mouse_y(void);
+float mouse_y(void);
 
 /// Returns mouse delta in X position between current and previous frames in
 /// screen coordinates.
 ///
 /// @returns Mouse delta in X position in screen coordinates.
 ///
-int mouse_dx(void);
+float mouse_dx(void);
 
 /// Returns mouse delta in Y position between current and previous frames in
 /// screen coordinates.
 ///
 /// @returns Mouse delta in Y position in screen coordinates.
 ///
-int mouse_dy(void);
+float mouse_dy(void);
 
 /// Checks whether a particular mouse button went down in the current frame.
 ///
